### 2023-03-01 #83 @teddygroves
bibat: a batteries-included Bayesian analysis template
Submitting Author: Teddy Groves (@teddygroves)
All current maintainers: @teddygroves
Package Name: bibat
One-Line Description of Package: An interactive template for Bayesian statistical analysis projects
Repository Link:  https://github.com/teddygroves/bibat/
Version submitted: 0.1.0
Archive: [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.8124312.svg)](https://doi.org/10.5281/zenodo.8124312)
Version accepted: 0.1.10
JOSS DOI:

---

Bibat is a Python package providing a flexible interactive template for Bayesian statistical analysis projects.

It aims to make it easier to create software projects that implement a Bayesian workflow that scales to arbitrarily many inter-related statistical models, data transformations, inferences and computations. Bibat also aims to promote software quality by providing a modular, automated and reproducible project that takes advantage of and integrates together the most up to date statistical software.

Bibat comes with "batteries included" in the sense that it creates a working example project, which the user can adapt so that it implements their desired analysis. We believe this style of template makes for better usability and easier testing of Bayesian workflow projects compared with the alternative approach of providing an incomplete skeleton project.

The target audience is people who want to write software implementing a Bayesian workflow as described in [this paper](https://arxiv.org/abs/2011.01808v1) and are willing to do so using Python scientific libraries, Stan, cmdstanpy, pydantic, pandera and make, as well as perhaps pytest, sphinx, quarto and github actions.

I am not aware of any interactive template that specifically targets a Python-based Bayesian workflow that scales to arbitrarily many models and data transformations.

In addition, bibat is unusual compared to other interactive templates because it is 'batteries-inclued', providing a full working example project rather than an incomplete skeleton

  - If you made a pre-submission enquiry, please paste the link to the corresponding issue, forum post, or other discussion, or `@tag` the editor you contacted: <https://github.com/pyOpenSci/software-submission/issues/82>

- [x] The package is deposited in a long-term repository with the DOI: [10.5281/zenodo.7688421](https://doi.org/10.5281/zenodo.7688421)

NB I submitted an old version of this package to JOSS and it was judged to be out of scope in terms of substantial scholarly effort. See here for the relevant issue in the joss-reviews repository : https://github.com/openjournals/joss-reviews/issues/4760. The reviewers did not definitively assess whether the package was within JOSS's scope in terms of not being a "minor utility" but noted that they had a query about this.

I think that bibat in its current form is within JOSS's scope in both of these respects but I imagine they would want to  assess this separately.

[Templates]: https://www.pyopensci.org/software-peer-review/appendices/templates.html
