### 2023-03-21 #83 @NickleDave
Hi again @teddygroves overall I can tick off most of the boxes for editor checks.

There is one issue we do need you to address before we can proceed.

- [ ] Add API docs. Even though the main use of the package is through the CLI, you want to document the internal API for yourself, for users, and for potential contributors
  - Not strictly required but I would import your modules in bibat/__init__.py and then define an `__all__` so that it's clear what "belongs" to `bibat` vs other things you import, e.g. importlib. A bit confusing to do `>>> dir(bibat)` and see `importlib`.

I also ran into a couple of minor issues that I describe in comments below but those are just there for feedback.

Please get those API docs added so we can go ahead with the review! Thank you.

  - [x] The package imports properly into a standard Python environment `import package-name`.
    - I *am* able to import in Python 3.9 without fail  (but see comments below)
    - Might be worth adding a conda-forge package (see comments below)
  - yes we consider this in scope under the category of reproducibility. This tool helps standardize computional projects / research compendia that perform Bayesian analyses, IIUC only through Stan (related libraries like PyMC are not currently included AFAICT)
  - [ ] API documentation (documentation for your code's functions, classes, methods and attributes): this includes clearly written docstrings with variables defined using a standard docstring format. *We suggest using the [Numpy](https://numpydoc.readthedocs.io/en/latest/format.html#docstring-standard) docstring format*.
      - contributor covenant
  - [x] **Code of Conduct** The package has a `Code of Conduct` file.
    - MIT
- [x] **Automated tests** Package has a testing suite and is tested via GitHub actions or another Continuous Integration service.

---
---

-  As noted the package imports correctly in Python 3.9, but I'm surprised I was able to install in a Python 3.7 environment. I first tried to install in a Python 3.7 environment accidentally and got an error: `ModuleNotFoundError: No module named 'importlib.metadata'`, which of course makes sense since this was added after Python 3.7, but I would expect the `pip install bibat` to fail because of the `requires_python` option in setup.cfg. It shows up properly as >=3.9 on PyPI. Not sure if there's a way to fix this so `pip install bibat` fails in a Python 3.7 venv

- I noticed the docs page is very slow to load -- it hangs a while the first time I click and then I have to refresh. Not clear how you are hosting the docs--GitHub pages? Might be worth switching to ReadTheDocs?

### Other notes
- Relying on `make` means that `Windows` users will have a harder time using `bibat`
- Similarly `pip` installing `requirements.txt` is likely to lead to issues on some platforms, e.g. for me I could not run the first `make analysis` to test that setup worked (as in the vignette) on Mac OS, because of a conflict with SciPy:

```
$  make analysis
. .venv/bin/activate && (\
	  python -m pip install --upgrade pip; \
	  python -m pip install -r requirements.txt --quiet; \
	  install_cmdstan ; \
	)
Requirement already satisfied: pip in /Users/davidnicholson/opt/anaconda3/lib/python3.7/site-packages (21.3.1)
Collecting pip
  Using cached pip-23.0.1-py3-none-any.whl (2.1 MB)
Installing collected packages: pip
  Attempting uninstall: pip
    Found existing installation: pip 21.3.1
    Uninstalling pip-21.3.1:
      Successfully uninstalled pip-21.3.1
Successfully installed pip-23.0.1
ERROR: Cannot install -r requirements.txt (line 1) and scipy because these package versions have conflicting dependencies.
ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/topics/dependency-resolution/#dealing-with-dependency-conflicts
/bin/sh: install_cmdstan: command not found
```

- An alternative would be to use `nox` with `conda` environments, which might be more platform agnostic. And/or a workflow tool like `snakemake`. Nox can have its own bootstrapping issues (see comment near end of https://github.com/pyOpenSci/software-submission/issues/68#issuecomment-1426357532) but it is mostly platform-agnostic and lets one write pure Python.
