### 2023-03-21 #83 @teddygroves
Thanks a lot for the checks and suggestions and for finding a reviewer!

I will add API documentation and follow the `__init__.py` convention that you suggest as soon as possible, and also check if there is a way to fail sooner for unsupported Python versions.

The documentation page is hosted on readthedocs. I'll have a look and see if I've done anything that makes it slow to load.

I'll have to think a bit about your points re make and pip/requirements.txt, thanks for the pointers to alternatives.
