### 2023-03-27 teddygroves/bibat#34 @teddygroves
Whether/how to not use make
See comments here https://github.com/pyOpenSci/software-submission/issues/83#issuecomment-1478527099: the current  dependency on make is not ideal for windows users.

On the other hand, since cmdstanpy depends on make already maybe this is the least bad option.

Maybe snakemake and/or nox could be partly incorporated - worth looking at.
