### 2023-04-26 #83 @mjhajharia
Hi! Reviews are due in 2 weeks, here's the [template](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#peer-review-template) that you can post as a comment to this issue. Feel free to ask anything that might be ambiguous.
