### 2023-04-26 teddygroves/bibat#36 @OriolAbril
comments on the vignette
The link to `data from the 2006 Major league season` is not rendered correctly (6th paragraph in intro)

"After I answered the wizard’s questions bibat creted a new folder called baseball that looked like this:" answers should be in the vignette (save author names and email, or maybe even with those and using placeholder ones)

I don't understand the part about "Since this worked, I added a new makefile target for the raw data files:", the `.` at the beginning of the last line doesn't look right but more importantly, this is not used anywhere.

Naming of paths is inconsistent accross vignette. Given how there are multiple files with similar names, I would stick to paths relatives to the project home for everything. For example, right at the beginning, we start editing `src/prepared_data.py`, then we move onto `data_preparation_functions.py` (which lives inside `src` too) and right after to `prepare_data.py` (now in the project home).

"To check that all this works, we can run the script prepare_data.py manually or using make analysis." doesn't work with `make analysis` because tests still refer to example data. Moreover, when running `python prepare_data.py` I get a pydantic validation error.

stopped here for now
