### 2023-04-26 #83 @OriolAbril
- [x] As the reviewer I confirm that there are no conflicts of interest for me to review this work

  * **wasn't able to locate the dev instructions in the readme nor contributing guide**
  * **the section appears to be empty? https://bibat.readthedocs.io/en/latest/#command-line-interface**
  * **It might also be a good idea to link to the code of conduct from CONTRIBUTING.md and https://bibat.readthedocs.io/en/latest/#contributing**
  * **I saw some comments above when skimming the previous discussion. I don't understand why both setup.py and setup.cfg are needed, pyproject.toml alone should do the same job. The pyopensci packaging guide also recommends using pyproject.toml**

<details><summary>Example of working pyproject.toml python requirement</summary>

I maintain [xarray-einstats](https://github.com/arviz-devs/xarray-einstats) which has a minimum python required in its pyproject.toml file. Today I have executed the following commands:

```
❯ conda create -n test python==3.7
Retrieving notices: ...working... done
Collecting package metadata (current_repodata.json): done
Solving environment: failed with repodata from current_repodata.json, will retry with next repodata source.
Solving environment: done

==> WARNING: A newer version of conda exists. <==
  current version: 22.11.1
  latest version: 23.3.1

Please update conda by running

    $ conda update -n base -c defaults conda

Or to minimize the number of packages updated during conda update use

     conda install conda=23.3.1

## Package Plan ##

  environment location: /home/oriol/bin/miniconda3/envs/test

  added / updated specs:
    - python==3.7

The following packages will be downloaded:

    package                    |            build
    ---------------------------|-----------------
    libffi-3.2.1               |    he1b5a44_1007          47 KB  conda-forge
    openssl-1.0.2u             |       h516909a_0         3.2 MB  conda-forge
    pip-23.1.2                 |     pyhd8ed1ab_0         1.3 MB  conda-forge
    python-3.7.0               |    hd21baee_1006        31.5 MB  conda-forge
    readline-7.0               |    hf8c457e_1001         391 KB  conda-forge
    setuptools-67.7.2          |     pyhd8ed1ab_0         569 KB  conda-forge
    sqlite-3.28.0              |       h8b20d00_0         1.9 MB  conda-forge
    wheel-0.40.0               |     pyhd8ed1ab_0          54 KB  conda-forge
    ------------------------------------------------------------
                                           Total:        39.0 MB

The following NEW packages will be INSTALLED:

  _libgcc_mutex      conda-forge/linux-64::_libgcc_mutex-0.1-conda_forge
  _openmp_mutex      conda-forge/linux-64::_openmp_mutex-4.5-2_gnu
  bzip2              conda-forge/linux-64::bzip2-1.0.8-h7f98852_4
  ca-certificates    conda-forge/linux-64::ca-certificates-2022.12.7-ha878542_0
  libffi             conda-forge/linux-64::libffi-3.2.1-he1b5a44_1007
  libgcc-ng          conda-forge/linux-64::libgcc-ng-12.2.0-h65d4601_19
  libgomp            conda-forge/linux-64::libgomp-12.2.0-h65d4601_19
  libstdcxx-ng       conda-forge/linux-64::libstdcxx-ng-12.2.0-h46fd767_19
  libzlib            conda-forge/linux-64::libzlib-1.2.13-h166bdaf_4
  ncurses            conda-forge/linux-64::ncurses-6.3-h27087fc_1
  openssl            conda-forge/linux-64::openssl-1.0.2u-h516909a_0
  pip                conda-forge/noarch::pip-23.1.2-pyhd8ed1ab_0
  python             conda-forge/linux-64::python-3.7.0-hd21baee_1006
  readline           conda-forge/linux-64::readline-7.0-hf8c457e_1001
  setuptools         conda-forge/noarch::setuptools-67.7.2-pyhd8ed1ab_0
  sqlite             conda-forge/linux-64::sqlite-3.28.0-h8b20d00_0
  tk                 conda-forge/linux-64::tk-8.6.12-h27826a3_0
  wheel              conda-forge/noarch::wheel-0.40.0-pyhd8ed1ab_0
  xz                 conda-forge/linux-64::xz-5.2.6-h166bdaf_0
  zlib               conda-forge/linux-64::zlib-1.2.13-h166bdaf_4

Proceed ([y]/n)? y

Downloading and Extracting Packages

Preparing transaction: done
Verifying transaction: done
Executing transaction: done
#
# To activate this environment, use
#
#     $ conda activate test
#
# To deactivate an active environment, use
#
#     $ conda deactivate

❯ conda activate test
❯ pip install xarray-einstats==0.5.1
ERROR: Ignored the following versions that require a different python version: 0.3.0 Requires-Python >=3.8; 0.4.0 Requires-Python >=3.8; 0.5.0 Requires-Python >=3.8; 0.5.1 Requires-Python >=3.8
ERROR: Could not find a version that satisfies the requirement xarray-einstats==0.5.1 (from versions: 0.1, 0.2.0, 0.2.1, 0.2.2)
ERROR: No matching distribution found for xarray-einstats==0.5.1
```

I am not a packaging expert, but I remember ipython had an installation bug not long ago due to combining multiple installation files. Using only pyproject.toml might simplify things and work like it does with xarray-einstats.

    - [x] Preferred citation
  * All points below are in the README, but in a different order
  * I am not aware of any similar package, so I think this is not necessary as of now, still wanted to check before deleting the check point
  * Extended citation info is missing, there is only the zenodo badge

  * **currently attempting to run the vignette, see https://github.com/teddygroves/bibat/issues/36 for details and progress**

Estimated hours spent reviewing: 26/04 -> 1h 40' (significant time spent going over reviewing guidelines too)

---
