### 2023-04-26 teddygroves/bibat#36 @teddygroves
Thanks very much for the comments and for catching some problems with the vignette.

The path naming should clearly be made consistent - sorry if it was confusing to go through! I also think it is kind of annoying to have all three of `prepare_data.py`, `src/prepared_data.py` and `src/data_prepararation_functions.py` in the first place and will try and consolidate these files.

I'll add the exact answers to the questions to the vignette. This is the reason for the test error you hit - I answered 'no' to the tests question but didn't record this.

The `.` in the makefile is to source the virtual environment as in https://unix.stackexchange.com/questions/114300/whats-the-meaning-of-a-dot-before-a-command-in-shell. The reason for adding the rule was to allow for automating fetching the raw data. However as you point out this isn't finished.

If you still have it could you possibly copy the pydantic error here?
