### 2023-04-27 teddygroves/bibat#36 @teddygroves
The pull request has been merged and I think the vignette and the experience as a whole are improved a bit.

I checked all of the paths in the vignette and added the answers to the questions as suggested.

I also changed the target project's structure so that all python files other than tests are in the same folder and there are only two data preparation files - `<package>/data_preparation.py` and `<package>/prepare_data.py`, with the latter being a thin script. I think finding data preparation logic is now a lot more intuitive.

Thanks again for the helpful comments - please let me know if you think they have been addressed!
