### 2023-04-27 #83 @teddygroves
Just recording here that in the latest version of bibat (0.1.4) some of the comments above are now addressed.

Specifically:

- readthedocs page now shows [information about the api and command line interface](https://bibat.readthedocs.io/en/latest/#command-line-interface)
- setup.py and setup.cfg replaced by pyproject.toml
- readme.rst now has all required badges except test coverage: I was just wondering if anyone has a recommendation as there seem to be quite a few options.
- bibtex citation info provided

Regarding dev installation instructions: there are instructions for pip installing the latest commit from the main branch, but there is no separate dev branch so that isn't linked. I'm not sure what is the best practice here but I'm very open to recommendations!
