### 2023-05-02 #83 @alizma
## Bibat Package Review

- [x] **Installation instructions:** for the development version of the
    **Comment**: Sections for The CLI, Python functions, and Wizarding module are incomplete on the official ReadTheDocs.

    **Comment**: CmdStan has a history of requiring extra effort to install for some users. Since it is a dependency here, including information about CmdStan's requirements would be a good idea (specifically the C++ compiler and make versions). Maybe linking to or drawing inspiration from the [official CmdStan docs](https://mc-stan.org/docs/cmdstan-guide/cmdstan-installation.html#cpp-toolchain) would be helpful.

Estimated hours spent reviewing: ~4

---

Overall, this package shows a lot of promise as a foundation for statistical analysis codes and provides a reusable, extendable base for the Bayesian workflow. Moreover, the utilities for generating reports are very nice. Here are a few general comments and things which could be improved in addition to other issue-specific comments above:

* For the sake of consistency and to reduce the barrier of entry for potential contributors, adding a ``pre-commit`` to be used in a developer environment would be useful. This requires the addition of a single file which sets up local checks prior to a commit and ensures that everything locally matches with CI requirements.
* It's not clear which versions of Python are usable. Python 3.9 works fine but some users might prefer to use a separate version. Are there any version-critical dependencies? More on this point, this would also involve extending the existing CI to confirming that builds work for other Python and dependency version combinations.
* The additional examples you provide that are based on the code that ``Bibat`` generates will be very useful for users, and you give a great example of transforming the set up code to a custom scenario.
* When first running ``investigate.ipynb``, it doesn't seem that ``arviz`` is being shipped as a requirement, meaning the default notebook cannot be run on first set-up.

###### Documentation

The current documentation is what seems to require the most attention. Specifically, some sections are entirely incomplete (as mentioned earlier) and a generic documentation for the generated environment are not provided. It suffices to automatically generate these for the default environment and include them on the overall documentation page. Additionally, the documentation should be split up into separate pages. For cleanliness, having a separate page for more details on the CLI, one for the default provided functions, the editing example you provided, and so on. Additionally, given the tools for automatically generating reports, it would be nice to have a generic example for doing this, especially for people who might not be familiar with using these tools.

The comments above with regards to default function documentation on the main documentation page apply equally well to the default documentation. Since projects based on this set-up will likely have problem-specific commentaries on each function, this is an additional important feature.

You also seem to be missing a linter check in your tests (use black or something similar). Also I'm not too sure how to assess the test coverage in your CI, but maybe you could check a few other things besides running the commands for the default example?
