### 2023-05-03 teddygroves/bibat#38 @teddygroves
split documentation into different pages
see comments [here](https://github.com/pyOpenSci/software-submission/issues/83#issuecomment-1532168928).
