### 2023-05-03 teddygroves/bibat#39 @teddygroves
Add example of automatically generating a report to the documentation
See comment [here](https://github.com/pyOpenSci/software-submission/issues/83#issuecomment-1532168928).
