### 2023-05-03 teddygroves/bibat#40 @teddygroves
Add a pre-commit
https://github.com/pyOpenSci/software-submission/issues/83#issuecomment-1532168928
