### 2023-05-03 teddygroves/bibat#41 @teddygroves
Rationalise minimum python version
https://github.com/pyOpenSci/software-submission/issues/83#issuecomment-1532168928

- [ ] Is it really necessary to have 3.9 as the minimum version?
- [ ] Make CI test all supported Python versions
- [ ] Make clear in documentation which versions are supported
