### 2023-05-03 teddygroves/bibat#42 @teddygroves
Check test coverage
https://github.com/pyOpenSci/software-submission/issues/83#issuecomment-1532168928

- [x] coverage tool
- [x] badge
- [x] check some more things
