### 2023-05-03 teddygroves/bibat#43 @teddygroves
Better documentation in the target project
https://github.com/pyOpenSci/software-submission/issues/83#issuecomment-1532168928
