### 2023-05-03 teddygroves/bibat#47 @teddygroves
Document dependencies
See https://github.com/pyOpenSci/software-submission/issues/83#issuecomment-1532168928

These are the ones likely to cause problems (cmdstan is generally troublesome and quarto is the only dependency apart from python and make that bibat doesn't try to install):

- [ ] cmdstan
- [ ] quarto
