### 2023-05-08 teddygroves/bibat#41 @teddygroves
After having a look through the repository and trawling my memory I have found the reason for having python 3.9 as the minimum version: some parts of the template project use the `|` dictionary union operator that was introduced in version 3.9.

There aren't that many places where the new operator is used---I think it is only in [this file](https://github.com/teddygroves/bibat/blob/main/bibat/%7B%7Bcookiecutter.repo_name%7D%7D/%7B%7Bcookiecutter.repo_name%7D%7D/sample.py)---so it wouldn't be too hard to make bibat compatible with earlier Python versions by reverting to the old syntaxt with `{**d1, **d2}`. However I'd rather avoid this as I like the new syntax.

Another consideration is that some best practices for using pydantic are different depending on the python version and I would rather stay up to date.
