### 2023-05-08 teddygroves/bibat#41 @teddygroves
Closing as documentation now makes clear which python versions are supported and the end to end test now runs on all supported python versions (the unit tests already did).
