### 2023-05-08 #83 @teddygroves
Thanks very much for the comments @alizma! I think they are now just about all addressed, and the package is overall a lot better as a result.

Pre-commit is a much nicer solution for linting than what I had previously.

Test coverage is now displayed in a readme badge.

Information about installing Cmdstan, as well as some lists of other dependencies, has been added to the readme and documentation.

The documentation has been split into multiple pages and generally improved. In particular, it is now explained at greater length how to document a bibat project, and the example Sphinx documentation includes automatic documentation generation for the `data_preparation` module.

A general problem that I don't really know what to do about is documenting and testing the example project code. Due to use of templates it generally isn't valid code, which makes automation difficult. What I've done is check that the example project runs from end to end without any errors, but there is definitely more that could be done.

Please let me know what you think, especially about the last point!
