### 2023-05-15 teddygroves/bibat#36 @OriolAbril
Tried running the vignette from scratch and I am now somehow having cmdstan installation issues. I am on Debian 11. Here is the error message:

```
Installing CmdStan version: 2.32.1
Install directory: /home/oriol/.cmdstan
Downloading CmdStan version 2.32.1
Download successful, file: /tmp/tmp0mrlhx_s
Extracting distribution
Unpacked download as cmdstan-2.32.1
Building version cmdstan-2.32.1, may take several minutes, depending on your system.
Traceback (most recent call last):
  File "/home/oriol/Documents/repos_oss/baseball/.venv/lib/python3.10/site-packages/cmdstanpy/install_cmdstan.py", line 282, in build
    do_command(cmd, fd_out=None)
  File "/home/oriol/Documents/repos_oss/baseball/.venv/lib/python3.10/site-packages/cmdstanpy/utils/command.py", line 76, in do_command
    raise RuntimeError(msg)
RuntimeError: Command ['make', 'build', '-j1']
	error during processing No such file or directory

Traceback (most recent call last):
  File "/home/oriol/Documents/repos_oss/baseball/.venv/bin/install_cmdstan", line 8, in <module>
    sys.exit(__main__())
  File "/home/oriol/Documents/repos_oss/baseball/.venv/lib/python3.10/site-packages/cmdstanpy/install_cmdstan.py", line 668, in __main__
    run_install(InstallationSettings(**args))
  File "/home/oriol/Documents/repos_oss/baseball/.venv/lib/python3.10/site-packages/cmdstanpy/install_cmdstan.py", line 602, in run_install
    install_version(
  File "/home/oriol/Documents/repos_oss/baseball/.venv/lib/python3.10/site-packages/cmdstanpy/install_cmdstan.py", line 422, in install_version
    build(verbose, progress=progress, cores=cores)
  File "/home/oriol/Documents/repos_oss/baseball/.venv/lib/python3.10/site-packages/cmdstanpy/install_cmdstan.py", line 286, in build
    raise CmdStanInstallError(f'Command "make build" failed\n{str(e)}')
cmdstanpy.install_cmdstan.CmdStanInstallError: Command "make build" failed
Command ['make', 'build', '-j1']
	error during processing No such file or directory
make: *** [Makefile:18: env] Error 1
```

I am afraid I don't understand the message and I'm not sure how to go about it. Sorry about the slow response too
