### 2023-05-15 teddygroves/bibat#36 @teddygroves
Thanks for trying again!

Installing cmdstan can be a pain unfortunately. From the error message I think it downloaded the cmdstan repository and saved it in the right place but the `make build -j1` command failed.

If you have time and the folder is still around it would be really helpful if you could try running `make build -j1` from the directory `/home/oriol/.cmdstan/cmdstan-2.32.1`. This should hopefully produce a more verbose error message that narrows down the problem a little.
