### 2023-05-15 teddygroves/bibat#36 @OriolAbril
Got this:

```
❯ make build -j1
stan/lib/stan_math/make/libraries:117: *** "Need to set TBB_CXX_TYPE for non-standard compiler other than gcc or clang.".  Stop.
```

which was already mentioned before, but I do have gcc available:

```
❯ gcc --version
gcc (conda-forge gcc 10.4.0-19) 10.4.0
Copyright (C) 2020 Free Software Foundation, Inc.
This is free software; see the source for copying conditions.  There is NO
warranty; not even for MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
```

(I even have cmdstanpy installed via conda forge and working on that same env)
