### 2023-05-15 teddygroves/bibat#36 @teddygroves
I’m a bit stumped: my best guess is that something has gone wrong due to conda-forge gcc doing something that cmdstan doesn’t expect. Specifically, I expect that in [here](https://github.com/stan-dev/math/blob/133fd015ea0389287e45d6688075eaa77cecd23d/make/compiler_flags) the variable `CXX_TYPE` is ‘other’ when it should be ’gcc’, leading to `TBB_CXX_TYPE` being wrong [here](https://github.com/stan-dev/math/blob/3789de62bab3b77b1009709554b5ac97ac5a54a3/make/libraries).

For a workaround I think the best thing to do is look at how conda-forge installs cmdstan. I’ll check that as soon as possible and post again here.

More generally I need to check which conda/non-conda combinations the Stan ecosystem supports and adjust bibat accordingly. Currently bibat doesn’t do anything conda related. This is partly because I haven’t got much experience with conda and partly because it seems like it would be complicated to support both conda and standard Python virtual environments, which I definitely want to keep. I’m very curious to hear what you think about the second consideration!
