### 2023-05-16 teddygroves/bibat#36 @teddygroves
I found a couple of reports about the same issue [here](https://github.com/stan-dev/cmdstan/issues/947) and [here](https://discourse.mc-stan.org/t/error-installing-cmdstan/19216/2). In both cases cmdstan failed to install with the same error and there was some conda stuff around.

Apparently adding these lines to the (possibly new) file `/home/oriol/.cmdstan/cmdstan-2.32.1/make/local` might work:

```
CXX=g++
TBB_CXX_TYPE=gcc
```

From [this file](https://github.com/conda-forge/cmdstan-feedstock/blob/447155668bd3fbef8df520dc0e98a7a3f69ad60c/recipe/build.sh) it looks like the conda-forge recipe for building cmdstan also needs to set the variable `TBB_CXX_TYPE` in `make/local`.
