### 2023-05-19 teddygroves/bibat#36 @OriolAbril
Restarted with a clean conda env (which had no gcc compiler so the system one was used and everything worked well). Here are some more comments.

* Somehow this line https://github.com/teddygroves/bibat/blob/main/bibat/%7B%7Bcookiecutter.repo_name%7D%7D/Makefile#L38 made it to my make analysis despite answering no to the question. pytest wasn't installed and there was no pytest folder (as expected) so that ended up in an error.
* I see now `PreparedData` class and similars are in `data_preparation.py` but the [vignette](https://bibat.readthedocs.io/en/latest/_static/report.html#preparing-the-data) has not been updated and still has `prepared_data.py`.
* Getting `prepare_data.py` to run requires a few extra steps than the ones that appear in the vignette. I deleted everything but the write_prepared_data one, copied the provided functions and then modified `prepare_data` function to:

  ```
  def prepare_data():
    """Run main function."""
    print("Reading raw data...")
    raw_data = {
        k: pd.read_csv(v, index_col=None) for k, v in RAW_DATA_FILES.items()
    }
    print("Preparing data...")
    print(f"Running data preparation 2006")
    prepared_data = prepare_data_2006(raw_data["2006"])
    output_dir = os.path.join(PREPARED_DIR, prepared_data.name)
    print(f"\twriting files to {output_dir}")
    if not os.path.exists(PREPARED_DIR):
        os.mkdir(PREPARED_DIR)
    write_prepared_data(prepared_data, output_dir)
    print(f"Running data preparation bdb")
    prepared_data = prepare_data_bdb(raw_data["bdb-main"], raw_data["bdb-post"], raw_data["bdb-apps"])
    output_dir = os.path.join(PREPARED_DIR, prepared_data.name)
    print(f"\twriting files to {output_dir}")
    if not os.path.exists(PREPARED_DIR):
        os.mkdir(PREPARED_DIR)
    write_prepared_data(prepared_data, output_dir)
  ```

* there is a note on the specifying inferences section that looks badly formatted
* after specifying inferences I realized `load_prepared_data` is indeed used, so I copied it back in. I don't think I added any coords nor dims though
* 2nd make analysis attempt ended up in import errors, I didn't realize I had to add the numpy and logit imports to the stan_input functions. numpy import is nearly everywhere, but depending on the python knowledge you expect users to have I'd mention the need to import `from scipy.special import logit`.
* 3rd attempt ended up in an error trying to execute investigate.ipynb and loading unexisting files (I removed the example inference folders). From my prior and the vignette itself, I was kind of expecting I'd run the notebook manually _after_ make analysis runs successfully, not that it'd be executed as part of make analysis too
