### 2023-05-23 teddygroves/bibat#36 @teddygroves
Thanks a lot for persevering with the installation issues and for the new comments. I think the PR linked above should address them all.
