### 2023-05-23 teddygroves/bibat#36 @OriolAbril
There is still a wrong formatted link in the vignette intro:

    [data from the 2006 Major league season])https://github.com/stan-dev/example-models/blob/master/knitr/pool-binary-trials/baseball-hits-2006.csv)

As for code execution itself, now the exclusion of the pytest line leaves an empty line in [its position](https://github.com/teddygroves/bibat/blob/main/bibat/%7B%7Bcookiecutter.repo_name%7D%7D/Makefile#L38), which breaks bash syntax for me. I got:

```
. .venv/bin/activate && (\

/bin/sh: 2: Syntax error: end of file unexpected (expecting ")")
make: *** [Makefile:35: analysis] Error 2
```

Removing the empty line from the instructions in `make analysis` fixed the issue, no idea how to get the template to add the pytest line only if necessary and leaving no empty lines in either case though.

The last line calling `write_prepared_data()` is missing from the provided `prepare_data()` code sample, there is also an empty black line at the beginning of the block. It also looks like `MEASUREMENTS_FILE` should not be deleted as indicated in

> To finish off I deleted the unused global variables MEASUREMENTS_FILE, NEW_COLNAMES, DROPNA_COLS and DIMS

After that managed to run everything to completion.

---

Here are some a bit more general coments.

I am still very confused about the intended workflow. It is mentioned that you can run scripts manually or run `make analysis` in multiple intermediate steps, but up until the very end of the notebook `make analysis` won't run successfully. Are users expected to run `make analysis` in each partial step until it runs into an error?

I see the goal, and think this is necessary for reproducibility, but I'd welcome a bit more guidance on this. i.e. how much of a bad user am I being if I ran `python baseball/sample.py` instead of `make analysis` for sampling? Leaving `make analysis` only for the very last step?

I am also a bit surprised by the use of json to store the mcmc results. Why not zarr or netcdf even? zarr should generate files at least 2/3 times smaller (maybe more if compressing) and supports distributed reading "out of the box". I fear using json might prevent the template from scaling to bigger analysis
