### 2023-05-23 teddygroves/bibat#36 @teddygroves
Thanks for catching those careless mistakes - sorry! The ones you pointed out are now fixed on the latest main branch, including the makefile whitespace: I should have checked that.

The workflow is intended to be neutral between running python scripts individually and running `make analysis` up to an error. In practice running the python script it is quicker as it avoids running some parts of the analysis, but running `make analysis` is also nice as it can sometimes catch an unexpected problem, plus having a single entry point means you don't have to remember the names of all the scripts and that `execute` is the command for running a jupyter notebook like a script. I don't think I wrote all that down anywhere though - thanks for highlighting the issue!  It should definitely be clearer that seeing errors is part of the intended workflow, and I can see how the way the vignette uses both methods is confusing.

Regarding the format for storing `InferenceData`s, this reflects my use - netcdf was broken on arm macs for a long period, and I didn't have a problem with too much output so I switched to json. I now quite like being able to read, copy bits of and even sometimes directly edit the json files. On the other hand as you say json is less scalable than the other formats so I guess they would be better in a lot of cases. I've made a new issue (#57) for looking again about this.
