### 2023-05-24 #83 @OriolAbril
Checked everything on my list after managing to run the vignette successfully. I recommend approving the package (starting from version >0.1.5, that is, from the next version based on current status of `main` branch).

As it has already been commented, testing can be extended, adding a 2nd case of end-to-end workflow for example, but I think it is good enough already. Given the inherent interactivity of the library and workflow, my recommendation would go more along the lines of aiming to rerun one of the vignettes manually before each minor release. (note I work mainly with visualization, so I am probably biased towards keeping a human in the loop and not relying 100% on automated checks)
