### 2023-05-24 teddygroves/bibat#36 @OriolAbril
Perfect, commented on the pyopensci issue too already. I'll close this now
