### 2023-05-30 #83 @mjhajharia
Hi! Firstly, thanks to @teddygroves for the great addition to software reviewed by PyOpenSci. Also, thanks a ton @alizma and @OriolAbril for taking the time to review this and suggest helpful improvements. Here are my informal reviewer 3 notes, most of these are optional, I hope you find them helpful, but they're mostly not big enough to be strict hindrances.

- I would recommend mentioning the requirements.txt or preferred python environment and such things for contributing in the contributing guide, it's not a major thing, but a good practice, not super relevant in this case as I imagine right now most people trying to contribute would have some experience because of this being a niche package, but it's nice for making it easier for newcomers as a standard practice. The tutorial on that page seems to be slightly confounding as it is a mix of using bibat for your own project versus actually contributing to core bibat code. I would also recommend conda because cmdstanpy can sometimes be a nuisance, but generally be ambivalent to provide free choice to users ofcourse!

- [this](https://bibat.readthedocs.io/en/latest/_static/report.html) looks very nice!
- [Code of conduct](https://github.com/teddygroves/bibat/CODE_OF_CONDUCT.md) link is broken.
- I recommend a separate page for CLI documentation instead of putting it on the [Reference](https://bibat.readthedocs.io/en/latest/reference.html) page.
- There's a broken link to [this](https://github.com/teddygroves/bibat/bibat/examples/baseball) URL on the [vignettes]([vignettes](https://bibat.readthedocs.io/en/latest/working_with_bibat.html#vignettes)) page, similar link [here](https://github.com/teddygroves/bibat/docs/index.rst)

- Generally, I think the documentation still feels somewhat unclear for somebody who isn't completely invested in what the package does already, for example - `prior`, `posterior` modes. I understand the purpose but even looking at the code it's slightly unclear what do they exactly do within the scripts. The example report is very nice, I just think some more details about the API would help a lot!
-
#### Optional nits regarding typos
- [Documenting your analysis](https://bibat.readthedocs.io/en/latest/working_with_bibat.html#documenting-your-analysis) has a typo in the line "Note that, unlike, bibat does not attempt to install or set up quarto: see [quarto’s ‘getting started’ page](https://quarto.org/docs/get-started/) for official installation instructions."

- There are a couple of proofreading issues in the documentation, which isn't a big deal, but in case you want to submit it to other places as well, it might be nice to have a quick run through the writing.

Accessibility

I recommend using an accessibility checker on the website and looking through some issues from this report for hints

    Color Contrast
![image](https://github.com/pyOpenSci/software-submission/assets/55153708/c8d011ed-af49-4a4c-aa96-b2e22376b73a)

    Improper use of preformatted text element
![image](https://github.com/pyOpenSci/software-submission/assets/55153708/d0cd8586-5b01-4803-a88f-126df05088cd)

    Line Height
    The line-height for paragraphs should be at least 1.5 times the font size.
![image](https://github.com/pyOpenSci/software-submission/assets/55153708/91e2f950-a60c-4953-8b5f-54ff053fd7b3)

    URL colours
![image](https://github.com/pyOpenSci/software-submission/assets/55153708/9126d232-8046-4ba4-87df-d16f786876ab)

![image](https://github.com/pyOpenSci/software-submission/assets/55153708/7fa66774-ebd0-417c-834d-646960bfd1c6)
![image](https://github.com/pyOpenSci/software-submission/assets/55153708/e88017d0-986b-4832-bfed-796854a6176b)

https://bibat.readthedocs.io/en/latest/_static/report.html
