### 2023-05-31 #83 @teddygroves
Thanks @mjhajharia for the latest comments. This review process has been a really fantastic learning experience for me!

The ['fix broken links' PR](https://github.com/teddygroves/bibat/pull/58) above addresses some of the typos and broken links.

Regarding the modes, I think this part of bibat is hard to understand at least partly because the underlying code isn't the best. I've started to address this in the ['Refactor to add...' PR](https://github.com/teddygroves/bibat/pull/59). The change requires updating the documentation and vignette so it will take a bit longer to arrive but should definitely be worth it.

Thanks for suggesting using an accessibility checker - I hadn't thought of doing this. Could you possibly tell me which one you used to generate the image above (or which one you recommend in general)?
