### 2023-06-02 teddygroves/bibat#60 @teddygroves
Monitor and improve documentation page accessibility
Use an accessibility checker as recommended [here](https://github.com/pyOpenSci/software-submission/issues/83#issuecomment-1568163887).
