### 2023-06-02 #83 @teddygroves
I've now done most of the things recommended, including:

- clarified the contributing instructions
- fixed broken readme and vignette links
- put cli and api docs on different pages
- proofread the documentation
- redesigned and redocumented fitting modes to make them more intuitive and extendable.

That leaves improving the documentation page's accessibility with a checker. I've made an issue for this and will start working on it as soon as possible.
