### 2023-06-06 teddygroves/bibat#60 @teddygroves
Checker added in #61 better theme added in #63
