### 2023-06-06 #83 @teddygroves
I've done a few things to improve the accessibility of bibat's documentation:

- added automatic accessibility auditing with lighthouse and AccessLint
- switched the documentation website to a new theme that scores better in the lighthouse audit and qualitatively seems like it has better contrast.
- added an accessibility page to the documentation following [this guide](https://www.gov.uk/government/publications/sample-accessibility-statement/sample-accessibility-statement-for-a-fictional-public-sector-website)

I've closed issue #60 from above as I think there has been an improvement but I'd still love to know the checker that @mjhajharia used as I think it noticed some things that the ones I added misssed.
