### 2023-06-21 #83 @teddygroves
Just checking here if there is something that someone is waiting for me to do.

I think almost all of the reviewers' major concerns have now been addressed. There are still [some live issues](https://github.com/teddygroves/bibat/issues) that arose from the reviews but I think they are relatively minor. If any of these are particularly pressing I'd appreciate a pointer so I can prioritise fixing them!
