### 2023-07-02 #83 @mjhajharia
Hi @lwasser, thanks for the ping, and apologies for the delay. Hi @teddygroves, thank you for the PRs I went through them and everything looks great, this is really fantastic and I really appreciate you taking time to incorporate the optional feedback!

> Thanks for suggesting using an accessibility checker - I hadn't thought of doing this. Could you possibly tell me which one you used to generate the image above (or which one you recommend in general)?

I think you're good to go in terms of accessibility now! For that report I used SiteImprove Accessibility checker extension, I'm not sure it's the best out there but it's easy to use and get a quick overview and points out major things in case this is helpful for you later!

In general though, if you follow standard accessibility guidelines that is a good place to be in for documentation, so thanks a ton for doing that!

Again sorry for the delay, had some personal emergencies, but this looks good to go as far as I am concerned, all the necessary things are resolved and the documentation looks. I am very happy to see this [page](https://bibat.readthedocs.io/en/latest/accessibility.html#accessibility), I've been meaning to make a PR in PyOpenSci docs about this as well and I think it would be super useful.

So this is good to go as far as I'm concerned, thankyou for the fantastic project @teddygroves!
