### 2023-07-05 #83 @lwasser
hey there @teddygroves !! 👋 congratulations on bibat being accepted into the pyOpenSci ecosystem!
I'm going to help wrap things up here. there are a few last things that need to happen!
And now for the formal acceptance proceedings
---

🎉  bibat has been approved by pyOpenSci! Thank you @teddygroves  for submitting bibat and many thanks to  @OriolAbril
and @alizma<reviewer-names-here> for reviewing this package! 😸

- [ ] Tag and create a release to create a Zenodo version and DOI. Please let me (@lwasser ) know when you do this so i can add the archive link to the top of this review!
- [ ] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.
- [ ] Add bibat to the pyOpenSci website. @teddygroves , please open a pr to update [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/packages.yml): to add your package and if you wish, add your name to the [list of contributors](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/contributors.yml) (i'm working on automating this but it's not quite done yet!).
- [ ] <reviewers-and-maintainers> if you have time and are open to being listed on our website, please add yourselves to [this file](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/contributors.yml) via a pr so we can list you on our website as contributors!

Finally - Teddy, if you are interested - we welcome a blog post that showcases bibat's functionality. this helps us promote your great work supporting the scientific Python ecosystem. This is of course optional but if you wish to do this - you can check out a few of our other blogs posts including: [pandera](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_posts/2019-11-18-pandera-dataframe-validation.md) and [moving pandas](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_posts/2020-04-29-movingpandas-movement-data-analysis.md) that have been published. these blogs are some of our most viewed content on the website!

---
