### 2023-07-07 #83 @teddygroves
Hi @lwasser, that's great news!

I think I've now done all the steps. The zenondo link for bibat version 0.1.10 is https://zenodo.org/record/8124312. I've added the badge and filled in the survey. The pr for updating packages.yml and contributors.yml is here: https://github.com/pyOpenSci/pyopensci.github.io/pull/216. I didn't know how to fill in some of the fields, e.g. `github_image_id` - I guess we can discuss that under the pr if necessary.

I'd love to contribute a blog post - for that do I just submit a pr for a new file [here](https://github.com/pyOpenSci/pyopensci.github.io/tree/main/_posts)?

Thanks a lot @lwasser, @mjhajharia @OriolAbril and @alizma - I really appreciated your time and attention and enjoyed the review process!
