### 2023-07-10 #83 @lwasser
@teddygroves thank you! and welcome to pyopensci!! yes please submit your pr / blog post to that link - you can look at the [moving pandas](https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_posts/2020-04-29-movingpandas-movement-data-analysis.md) and pandera blogs to get a sense of what others have done. but there are no requirements! please just showcase what your awesome tool provides!

before this is closed i'll go through all of the checks above to ensure it's all done!! i hope to do that in the next 3 days - i'm traveling right now but will followup soon!
