### 2023-07-12 #83 @lwasser
ok @teddygroves i'm going through this all today. your package will be live on our website by the end of the day tomorrow at the latest!

thank you everyone for supporting this review!! @mjhajharia for leading the review as editor! @OriolAbril @alizma for your reviews and @teddygroves for submitting to us and doing the work via peer review!

all - if anyone would like to join our slack group please either email me at leah at pyopensci.org OR respond here if i can grab your email from your github profile!!  closing this issue now! ✨
