+++ docs/conf.py
@@ -24,7 +24,7 @@ copyright = "2022, Teddy Groves"
 author = "Teddy Groves"
 
 # The full version, including alpha/beta/rc tags
-release = "0.1.12"
+release = "0.1.13"
 
 root_doc = "index"
 
