### 2023-12-13 #83 @lwasser
this issue can be closed as JOSS did [deem this review out of scope](https://github.com/openjournals/joss-reviews/issues/5678) . Many thanks everyone for your work here on the pyOpenSci review side of things!! Everyone involved in this review is welcome to join our slack group. if you haven't received an invitation feel free to reach out to @kierisi !! closing this now!!
